Refactor checkValueInRange method
